+++ server/io.js
@@ -1,0 +1,10 @@
+var io = require('socket.io')();
+
+io.on('connection', function (socket) {
+    socket.emit('news', { hello: 'world' });
+    socket.on('my other event', function (data) {
+        console.log(data);
+    });
+});
+
+module.exports = io;