+++ CHANGELOG.md
@@ -99,3 +99,6 @@
 - Add ActionEdge to link two states
 - Add HttpClient package to help send http requests
 - Add HttpUtil
+
+### (2016/4/14)
+- Add Module Server written in node.js